Add class and interface for Login To API
